finished the home page, no more errors
